move *keyFn to hub/api, make pulpId more explicit
