+++ frontend/hub/namespaces/hooks/useDeleteHubNamespaces.tsx
@@ -1,10 +1,9 @@
 import { useTranslation } from 'react-i18next';
 import { compareStrings, useBulkConfirmation } from '../../../../framework';
 import { requestDelete } from '../../../common/crud/Data';
-import { nameKeyFn } from '../../usePulpView';
 import { HubNamespace } from '../HubNamespace';
 import { useHubNamespacesColumns } from './useHubNamespacesColumns';
-import { hubAPI } from '../../api';
+import { hubAPI, nameKeyFn } from '../../api';
 
 export function useDeleteHubNamespaces(onComplete: (namespaces: HubNamespace[]) => void) {
   const { t } = useTranslation();
