+++ frontend/hub/remote-registries/RemoteRegistries.tsx
@@ -1,12 +1,12 @@
 import { useTranslation } from 'react-i18next';
 import { PageHeader, PageLayout, PageTable } from '../../../framework';
-import { pulpHRefKeyFn, useHubView } from '../useHubView';
+import { useHubView } from '../useHubView';
 import { RemoteRegistry } from './RemoteRegistry';
 import { useRemoteRegistriesActions } from './hooks/useRemoteRegistriesActions';
 import { useRemoteRegistriesColumns } from './hooks/useRemoteRegistriesColumns';
 import { useRemoteRegistryActions } from './hooks/useRemoteRegistryActions';
 import { useRemoteRegistryFilters } from './hooks/useRemoteRegistryFilters';
-import { hubAPI } from '../api';
+import { hubAPI, pulpHrefKeyFn } from '../api';
 
 export function RemoteRegistries() {
   const { t } = useTranslation();
@@ -14,7 +14,7 @@ export function RemoteRegistries() {
   const tableColumns = useRemoteRegistriesColumns();
   const view = useHubView<RemoteRegistry>({
     url: hubAPI`/_ui/v1/execution-environments/registries/`,
-    keyFn: pulpHRefKeyFn,
+    keyFn: pulpHrefKeyFn,
     toolbarFilters,
     tableColumns,
   });
