+++ frontend/hub/repositories/Repositories.tsx
@@ -16,9 +16,9 @@ import {
   TextCell,
 } from '../../../framework';
 import { RouteObj } from '../../Routes';
-import { hubKeyFn, pulpHRefKeyFn, useHubView } from '../useHubView';
+import { useHubView } from '../useHubView';
 import { RemoteRepository, Repository } from './Repository';
-import { hubAPI } from '../api';
+import { hubAPI, pulpIdKeyFn, pulpHrefKeyFn } from '../api';
 
 export function Repositories() {
   const { t } = useTranslation();
@@ -48,7 +48,7 @@ export function LocalRepositories() {
   const tableColumns = useLocalRepositoriesColumns();
   const view = useHubView<Repository>({
     url: hubAPI`/_ui/v1/distributions/`,
-    keyFn: hubKeyFn,
+    keyFn: pulpIdKeyFn,
     tableColumns,
   });
   return (
@@ -102,7 +102,7 @@ export function RemoteRepositories() {
   const tableColumns = useRemoteRepositoriesColumns();
   const view = useHubView<RemoteRepository>({
     url: hubAPI`/_ui/v1/remotes/`,
-    keyFn: pulpHRefKeyFn,
+    keyFn: pulpHrefKeyFn,
     tableColumns,
   });
   const rowActions = useMemo<IPageAction<RemoteRepository>[]>(
